Split channel.rs into mpsc and ipc files
Fixes #573
